Allow to have attr on first <section> elements.
Attributes for `.. revealjs_section::` doesn't work on first section of a
group. For example, `data-background-color` doesn't work on the first slide,
the title one.

Also when you set a `data-background-color` under header level 2, the
background apply to the bounding <section>, so all sub sections get this
background color.

Now, attributes are set to the first no-bounding <section>. So
`data-background-color` works on the title slide and it applies on the header
level 2 slides.